# Improve performance of list-of implementation