--- a/NEWS.md
+++ b/NEWS.md
@@ -2,6 +2,9 @@
 
 * Improved the performance of list-of common type methods (#1686).
 
+* The list-of method for `as_list_of()` now places the optional `.ptype`
+  argument after the `...` (#1686).
+
 * `vec_rbind()` now applies `base::c()` fallback recursively within
   packed df-cols (#1331, #1462, #1640).