A few improvements to the list-of implementation including:
- Avoiding calling `!!!` on a list-of. This has to go through `rlang:::rlang_as_list()` to unstructure the list, which can be expensive if called from the cast method repeatedly
- Faster internal `new_list_of0()` that avoids a few checks that don't need when passing lists around internally
- Avoiding dispatch of `as_list_of()` in the self-self cast method

A benchmark using a simplified version of the example from #1496 (this only focuses on the list-of part, not the list-col-in-a-tibble part)

``` r
library(vctrs)

make_list_of <- function(n) {
  vec_chop(new_list_of(vec_chop(1:n), ptype = integer()))
}

list1 <- make_list_of(1e3)
list2 <- make_list_of(2e3)
list4 <- make_list_of(4e3)
list8 <- make_list_of(8e3)

ptype <- vec_ptype(list1[[1]])

bench::mark(
  list1 = list_unchop(list1, ptype = ptype),
  list2 = list_unchop(list2, ptype = ptype),
  list4 = list_unchop(list4, ptype = ptype),
  list8 = list_unchop(list8, ptype = ptype),
  min_iterations = 200,
  check = FALSE
)

# Main
#> Warning: Some expressions had a GC in every iteration; so filtering is disabled.
#> # A tibble: 4 × 6
#>   expression      min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr> <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 list1        39.2ms   44.6ms     22.1     94.6KB     16.7
#> 2 list2        64.6ms   83.1ms     12.1     70.5KB     18.3
#> 3 list4       131.8ms    146ms      6.58   140.9KB     20.0
#> 4 list8       277.7ms  350.7ms      2.86   281.5KB     17.4

# This PR
#> Warning: Some expressions had a GC in every iteration; so filtering is disabled.
#> # A tibble: 4 × 6
#>   expression      min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr> <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 list1        24.2ms   28.9ms     34.0     75.3KB     17.9
#> 2 list2        49.4ms   53.4ms     18.4     70.5KB     19.4
#> 3 list4       100.1ms  105.8ms      9.27   140.9KB     19.4
#> 4 list8       202.6ms    214ms      4.62   281.5KB     19.4
```